requirements.txt
pyrogram==2.0.0
pymongo==4.1.0
python-dotenv==0.19.2
